Code/benchmarks-v2/requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# vLLM for text generation benchmarks
vllm>=0.2.0

# Data processing and visualization
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# System monitoring
psutil>=5.9.0

# Image processing
Pillow>=9.0.0

# Optional: for faster tokenization
tokenizers>=0.13.0
